Implement template equality
There are debug prints and TODOs that need to be resolved

With equality implemented and a hash function that returns a constant,
a set of unique templates can be formed. In the current state,
the 514 templates are reduced to 255 unique templates. The Kushman paper
says that there should only be 28.

One issue is that not all number slots are being parsed out correctly.
This is because sympy is multiplying numbers from the problem together
which results in numbers not in the problem text. The multiplied
numbers cannot be found in the text and so are left as constants