Move code from Template to NLP, begin derivations
To create a derivation, both the numbers and nouns from the NLP
are required. Template also needs to know what numbers are in the NLP
to determine if a value is a constant or not. Having both uses of
numbers go through the NLP class seemed best. This also allows adding
the nouns() function in the same spot.

The numbers and nouns are selected from the NLP
now just need to assign them to slots in all possible ways