Generate all possible derivations
Given a word problem and the set of unique templates to try,
the full set of possible derivations is being generated by trying
all templates and for each template trying all valid assignments of
nouns and numbers to slots.

From the Kushman paper:
  * nouns are allowed to map to multiple unknown slots
  * numbers can only map to a single number slot
    (but number slots can appear multiple times in a template)

The derivations use generator functions since having all possible derivations
in memory at once is not feasible. As an example:

  $python main.py extract-features -i 823

Found 27,897,120 derivations for problem 823 using the 30 templates