Update README with remaining tasks
The TODO section of the README includes all the tasks which I think
are necessary to have a functional implementation of what is described
in the Kushman paper. We do not need to implement all the features
or do everything exactly the way it is done in the paper. However, if we
do "something" for each task listed, then we should be able to measure
the performance of the system.

Having measurable performance will give us enough info to write a report.
Even if the performance is poor, we will be able to comment on what problems
it is able to correctly solve and what sort of mistakes it makes.